requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
autopep8==2.0.1
boto3==1.34.108
botocore==1.34.108
certifi==2022.6.15
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.1
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jmespath==1.0.1
openai==1.30.1
pillow==10.3.0
pycodestyle==2.10.0
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tomli==2.0.1
tqdm==4.66.4
typing_extensions==4.11.0
urllib3==2.2.1
